Improve handling of close and reload events (#10379)
- Allows Electron to handle exit asynchronously
- Allows browser & Electron to show different exit confirmations for different
  contributions
  Breaking:
    - canUnload() removed from WindowService interface
    - reload(), safeToShotDown() added to WindowService interface
    - polling of FrontentContributions moved to
      collectContributionUnloadVetoes()

+++ packages/core/src/browser/common-frontend-contribution.ts
@@ -20,7 +20,7 @@ import debounce = require('lodash.debounce');
 import { injectable, inject, optional } from 'inversify';
 import { MAIN_MENU_BAR, SETTINGS_MENU, MenuContribution, MenuModelRegistry, ACCOUNTS_MENU } from '../common/menu';
 import { KeybindingContribution, KeybindingRegistry } from './keybinding';
-import { FrontendApplication, FrontendApplicationContribution } from './frontend-application';
+import { FrontendApplication, FrontendApplicationContribution, OnWillStopAction } from './frontend-application';
 import { CommandContribution, CommandRegistry, Command } from '../common/command';
 import { UriAwareCommandHandler } from '../common/uri-command-handler';
 import { SelectionService } from '../common/selection-service';
@@ -55,6 +55,9 @@ import { QuickInputService, QuickPick, QuickPickItem } from './quick-input';
 import { AsyncLocalizationProvider } from '../common/i18n/localization';
 import { nls } from '../common/nls';
 import { CurrentWidgetCommandAdapter } from './shell/current-widget-command-adapter';
+import { ConfirmDialog, confirmExit, Dialog } from './dialogs';
+import { WindowService } from './window/window-service';
+import { FrontendApplicationConfigProvider } from './frontend-application-config-provider';
 
 export namespace CommonMenus {
 
@@ -352,6 +355,9 @@ export class CommonFrontendContribution implements FrontendApplicationContributi
 
     @inject(AuthenticationService)
     protected readonly authenticationService: AuthenticationService;
+
+    @inject(WindowService)
+    protected readonly windowService: WindowService;
 
     async configure(app: FrontendApplication): Promise<void> {
         const configDirUri = await this.environments.getConfigDirUri();
@@ -967,10 +973,10 @@ export class CommonFrontendContribution implements FrontendApplicationContributi
         });
     }
 
-    onWillStop(): true | undefined {
+    onWillStop(): OnWillStopAction | undefined {
         try {
             if (this.shouldPreventClose || this.shell.canSaveAll()) {
-                return true;
+                return { reason: 'Dirty editors present', action: () => confirmExit() };
             }
         } finally {
             this.shouldPreventClose = false;
@@ -983,10 +989,11 @@ export class CommonFrontendContribution implements FrontendApplicationContributi
         for (const additionalLanguage of ['en', ...availableLanguages]) {
             items.push({
                 label: additionalLanguage,
-                execute: () => {
-                    if (additionalLanguage !== nls.locale) {
+                execute: async () => {
+                    if (additionalLanguage !== nls.locale && await this.confirmRestart()) {
+                        this.windowService.setSafeToShutDown();
                         window.localStorage.setItem(nls.localeId, additionalLanguage);
-                        window.location.reload();
+                        this.windowService.reload();
                     }
                 }
             });
@@ -996,6 +1003,16 @@ export class CommonFrontendContribution implements FrontendApplicationContributi
                 placeholder: CommonCommands.CONFIGURE_DISPLAY_LANGUAGE.label,
                 activeItem: items.find(item => item.label === (nls.locale || 'en'))
             });
+    }
+
+    protected async confirmRestart(): Promise<boolean> {
+        const shouldRestart = await new ConfirmDialog({
+            title: nls.localizeByDefault('A restart is required for the change in display language to take effect.'),
+            msg: nls.localizeByDefault('Press the restart button to restart {0} and change the display language.', FrontendApplicationConfigProvider.get().applicationName),
+            ok: nls.localizeByDefault('Restart'),
+            cancel: Dialog.CANCEL,
+        }).open();
+        return shouldRestart === true;
     }
 
     protected selectIconTheme(): void {
